requirements.txt
pandas
s3fs
kafka-python